Refactor: Move email utility functions to utils.js
